Revert "feat: add GitHub Pages preview support for fork"
This reverts commit f277c4462215e59639b82896daef29e5b0724f4d.
