+++ next.config.ts
@@ -13,26 +13,12 @@ const cspHeader = `
     upgrade-insecure-requests;
 `;
 
-const isGitHubPagesBuild = process.env.GITHUB_PAGES === "true";
-const repositoryName = process.env.GITHUB_REPOSITORY?.split("/")[1] || "";
-const repositoryBasePath = isGitHubPagesBuild && repositoryName ? `/${repositoryName}` : "";
-
 const nextConfig: NextConfig = {
-  output: isGitHubPagesBuild ? "export" : "standalone",
-  basePath: repositoryBasePath,
-  assetPrefix: repositoryBasePath || undefined,
-  trailingSlash: isGitHubPagesBuild,
-  images: {
-    unoptimized: isGitHubPagesBuild,
-  },
+  output: "standalone",
   reactCompiler: true,
   poweredByHeader: false,
   allowedDevOrigins: ["trackerpathways.home.arpa"],
   async headers() {
-    if (isGitHubPagesBuild) {
-      return [];
-    }
-
     return [
       {
         source: "/:path*",
